user module 97% completed and primary table editing 100% completed panding editing user details 0%, user post delete 100% cmplt, tables block status added for admin

+++ sdCommunity/src/main/java/sdCommunity/user/profile/ProfileServlet.java
@@ -46,7 +46,7 @@ public class ProfileServlet extends HttpServlet {
 		String fname=(String)session.getAttribute("fname"); 
 		String lname=(String)session.getAttribute("lname"); 
 		String email=(String)session.getAttribute("email");
-		Integer num=(Integer)session.getAttribute("num");;
+		Long num=(Long)session.getAttribute("num");;
 		
 		String company = request.getParameter("company");
 		String about = request.getParameter("about");
